requirements.txt
python-telegram-bot==20.0
ccxt
pandas
pandas_ta
python-dotenv
openai
requests

